requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
bleach==6.3.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.4
colorama==0.4.6
crispy-bootstrap4==2025.6
cryptography==44.0.0
distro==1.9.0
Django==5.1.5
django-allauth==65.3.0
django-axes==8.1.0
django-cors-headers==4.9.0
django-crispy-forms==2.3
django-csp==4.0
django-filter==24.3
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.15.2
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.13.0
oauthlib==3.2.2
openai==2.24.0
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
pycparser==2.21
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.0.1
redis==7.2.1
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
sentry-sdk==1.40.0
setuptools==75.8.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.3.0
webencodings==0.5.1
whitenoise==6.6.0